+++ scripts/silver/envi_silver_layer/ddl_silver.sql
@@ -78,7 +78,7 @@ CREATE TABLE silver.envi_diesel_consumption (
     unit_of_measurement      VARCHAR(15),
     consumption              DOUBLE PRECISION,        -- Allows decimal values (e.g., 234.789)
     month                    VARCHAR(10),
-    year                     INT,
+    year                     SMALLINT,
     quarter                  VARCHAR(2),
 	date					 date,
     -- Metadata
@@ -96,7 +96,7 @@ CREATE TABLE silver.envi_electric_consumption (
     unit_of_measurement     VARCHAR(15),
     consumption             DOUBLE PRECISION,        -- Allows decimal values (e.g., 234.789)
     quarter                 VARCHAR(2),
-    year                    INT,
+    year                    SMALLINT,
     -- Metadata
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -114,7 +114,7 @@ CREATE TABLE silver.envi_non_hazard_waste (
     unit_of_measurement     VARCHAR(15),
     waste                   DOUBLE PRECISION,     -- Allows decimal values (e.g., 234.789)
     month                   VARCHAR(10),
-    year                    INT,
+    year                    SMALLINT,
     quarter                 VARCHAR(2),
     -- Metadata
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -133,7 +133,7 @@ CREATE TABLE silver.envi_hazard_waste_generated (
     unit_of_measurement     VARCHAR(15),
     waste_generated         DOUBLE PRECISION,                 -- Allows decimal values (e.g., 234.789)
     quarter                 VARCHAR(2),               -- Example: 'Q1', 'Q2', 'Q3', 'Q4'
-    year                    INT,
+    year                    SMALLINT,
     -- Metadata
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -149,7 +149,7 @@ CREATE TABLE silver.envi_hazard_waste_disposed (
     metrics                 VARCHAR(20),
     unit_of_measurement     VARCHAR(15),
     waste_disposed          DOUBLE PRECISION,     -- Allows decimal values (e.g., 234.789)
-    year                    INT,
+    year                    SMALLINT,
     -- Metadata
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
